requirements.txt
certifi==2020.4.5.2
chardet==3.0.4
Django==2.1
djangorestframework==3.11.0
httpie==2.1.0
idna==2.9
Pygments==2.6.1
pytz==2020.1
requests==2.23.0
sqlparse==0.3.1
urllib3==1.25.9
